Detect whether symbols actually change
There can lead to some redundant processing, but now the tool
can avoid writing a new file if nothing will change, and won't
print the file name (or highlights) when nothing did change.

This could be done up front in the tokenizer as an optimization.
The slight disadvantage of that would be leaking Converter
configuration into the Tokenizer.