Add wildcard matching and combine rules files
The replace/exclude list files are now combined into a single file
which uses '!' to indicate an exclusion. This reduces the number
of options and keeps flag soup a little more under control, and
feels like a more obvious and powerful way to specify behavior.

Wildcard matching is naive and has poor time complexity, but does
the job for now.