Add some crypto and OS names to lists
This exposed that the simple per-token on/off behavior is not always
sufficient for high-quality translation. Running this test on
std.crypto revealed several cases of phrases that should be preserved,
but are embedded in longer identifiers that should still be
otherwise updated. A slightly more powerful conversion rule system
is needed to accommodate these. It may be the last 5-10% that is
affected, which is a lot in a large code base like the Zig compiler.